No height respectively 0px
If you dont't want any height information for the background-image you may now set it to zero '0'. There might be a better solution with percental details, but I've just needed a quick fix and this works for me.
If this is approved, I may change the description within the settings of the plugin to give the user a hint of this feature.